+++ common/engine/keyboardprocessor/src/keyboard.cpp
@@ -10,41 +10,60 @@
 #include "keyboard.hpp"
 #include "json.hpp"
 #include "processor.hpp"
+#include "utfcodec.hpp"
 
 using namespace km::kbp;
 
-std::string utf16_to_utf8(std::u16string utf16_string);
 
-keyboard::keyboard(std::filesystem::path const & path)
-: _keyboard_id(path.stem().u16string()),
-  _version_string(u"3.145"),
-  _folder_path(path.parent_path()),
-  _default_opts {KM_KBP_OPTIONS_END}
+inline
+void keyboard_attributes::render()
 {
-  version_string = _version_string.c_str();
-  id = _keyboard_id.c_str();
-  folder_path = _folder_path.native().c_str();
-
-  if (path.extension() == ".kmx" ||
-      path.extension() == ".KMX") { // Some legacy packages may include upper-case file extensions
-    _processor = new kmx_processor(this);
-  }
-  else if (path.extension() == ".mock") {
-    _processor = new mock_processor(this);
-  }
-  else {
-    _processor = new null_processor(this);
-  }
-
+  // Make attributes point to the stored values above.
+  id              = _keyboard_id.c_str();
+  version_string  = _version_string.c_str();
+  folder_path     = _folder_path.c_str();
   default_options = _default_opts.data();
 }
 
-json & km::kbp::operator << (json & j, km::kbp::keyboard const & kb)
+
+keyboard_attributes::keyboard_attributes(std::u16string const & kbid,
+    std::u16string const & version,
+    path_type const & path,
+    options_store const &opts)
+: _keyboard_id(kbid),
+  _version_string(version),
+  _folder_path(path),
+  _default_opts(opts)
+{
+  // Ensure that the default_options array will be properly terminated.
+  _default_opts.push_back(option());
+  render();
+}
+
+
+keyboard_attributes::keyboard_attributes(keyboard_attributes &&rhs)
+: _keyboard_id(std::move(rhs._keyboard_id)),
+  _version_string(std::move(rhs._version_string)),
+  _folder_path(std::move(rhs._folder_path)),
+  _default_opts(std::move(rhs._default_opts))
+{
+    rhs.id = rhs.version_string = nullptr;
+    render();
+}
+
+
+keyboard_attributes & keyboard_attributes::operator = (keyboard_attributes &&rhs)
+{
+  return *new (this) keyboard_attributes(std::move(rhs));
+}
+
+
+json & km::kbp::operator << (json & j, km::kbp::keyboard_attributes const & kb)
 {
   j << json::object
-      << "id" << utf16_to_utf8(kb.id)
-      << "folder" << kb._folder_path.string()
-      << "version" << utf16_to_utf8(kb.version_string)
+      << "id" << std::u16string(kb.id)
+      << "folder" << kb._folder_path
+      << "version" << std::u16string(kb.version_string)
       << "rules" << json::array << json::close;
 
   return j << json::close;
@@ -56,21 +75,21 @@ json & km::kbp::operator << (json & j, km::kbp::keyboard const & kb)
   https://stackoverflow.com/a/35103224/1836776
 */
 
-#if _MSC_VER >= 1900 /* VS 2015 */ && _MSC_VER <= 1916 /* VS 2017 19.16 */
-
-std::string utf16_to_utf8(std::u16string utf16_string)
-{
-  std::wstring_convert<std::codecvt_utf8_utf16<int16_t>, int16_t> convert;
-  auto p = reinterpret_cast<const int16_t *>(utf16_string.data());
-  return convert.to_bytes(p, p + utf16_string.size());
-}
-
-#else
-
-std::string utf16_to_utf8(std::u16string utf16_string)
-{
-  std::wstring_convert<std::codecvt_utf8_utf16<char16_t>, char16_t> convert;
-  return convert.to_bytes(utf16_string);
-}
-
-#endif
+// #if _MSC_VER >= 1900 /* VS 2015 */ && _MSC_VER <= 1916 /* VS 2017 19.16 */
+//
+// std::string utf16_to_utf8(std::u16string utf16_string)
+// {
+//   std::wstring_convert<std::codecvt_utf8_utf16<int16_t>, int16_t> convert;
+//   auto p = reinterpret_cast<const int16_t *>(utf16_string.data());
+//   return convert.to_bytes(p, p + utf16_string.size());
+// }
+//
+// #else
+//
+// std::string utf16_to_utf8(std::u16string utf16_string)
+// {
+//   std::wstring_convert<std::codecvt_utf8_utf16<char16_t>, char16_t> convert;
+//   return convert.to_bytes(utf16_string);
+// }
+//
+// #endif