+++ common/engine/keyboardprocessor/src/keyboard.hpp
@@ -8,17 +8,13 @@
 
 #pragma once
 
-#include <experimental/filesystem>
 #include <string>
+#include <type_traits>
 
 #include <keyman/keyboardprocessor.h>
 
 #include "option.hpp"
-#include "processor.hpp"
-
-namespace std {
-  namespace filesystem = std::experimental::filesystem;
-}
+#include "path.hpp"
 
 // Forward declarations
 class json;
@@ -26,30 +22,41 @@ class json;
 namespace km {
 namespace kbp
 {
-
-  class keyboard : public km_kbp_keyboard_attrs
+  class keyboard_attributes : public km_kbp_keyboard_attrs
   {
-    std::u16string const                      _keyboard_id;
-    std::u16string const                      _version_string;
-    std::filesystem::path              const  _folder_path;
-    std::vector<km_kbp_option_item>           _default_opts;
-    abstract_processor                       *_processor;
+    std::u16string      _keyboard_id;
+    std::u16string      _version_string;
+    kbp::path           _folder_path;
+    std::vector<option> _default_opts;
+
+    void render();
+
   public:
-    keyboard(std::filesystem::path const &);
-    ~keyboard() {
-      delete _processor;
-    }
+    using options_store = decltype(_default_opts);
+    using path_type = decltype(_folder_path);
 
-    friend json & operator << (json &, km::kbp::keyboard const &);
+    keyboard_attributes()
+    : km_kbp_keyboard_attrs {nullptr, nullptr, nullptr, nullptr} {}
+    keyboard_attributes(keyboard_attributes const &) = delete;
+    keyboard_attributes(keyboard_attributes &&);
 
-    std::vector<km_kbp_option_item> *default_opts() { return & _default_opts; }
+    keyboard_attributes(std::u16string const & id,
+             std::u16string const & version,
+             path_type const & path,
+             options_store const &opts);
 
-    kbp::abstract_processor const &  processor() const noexcept { return *_processor; }
+    keyboard_attributes & operator = (keyboard_attributes const &) = delete;
+    keyboard_attributes & operator = (keyboard_attributes &&);
+
+    friend json & operator << (json &, km::kbp::keyboard_attributes const &);
+
+    options_store const &   default_opts_store() const noexcept { return _default_opts; }
+    options_store &         default_opts_store() noexcept { return _default_opts; }
+
+    path_type const     &   path() const noexcept { return _folder_path; }
   };
 
-  json & operator << (json &, km::kbp::keyboard const &);
+  json & operator << (json &, km::kbp::keyboard_attributes const &);
 
 } // namespace kbp
 } // namespace km
-
-struct km_kbp_keyboard : public km::kbp::keyboard {};