+++ common/engine/keyboardprocessor/src/km_kbp_keyboard_api.cpp
@@ -10,7 +10,6 @@
 */
 #include <cassert>
 #include <algorithm>
-#include <experimental/filesystem>
 #include <iterator>
 #include <sstream>
 #include <unordered_map>
@@ -18,37 +17,52 @@
 #include <vector>
 
 #include <keyman/keyboardprocessor.h>
+#include <kmx/kmx_processevent.hpp>
+#include <mock/mock_processor.hpp>
 #include <json.hpp>
 
 #include "keyboard.hpp"
 #include "option.hpp"
 
 
+using namespace km::kbp;
 
+namespace
+{
+  abstract_processor * processor_factory(path const & kb_path)
+  {
+    // Some legacy packages may include upper-case file extensions
+    if (kb_path.suffix() == ".kmx" || kb_path.suffix() == ".KMX") {
+      return new kmx_processor(kb_path);
+    }
+    else if (kb_path.suffix() == ".mock") {
+      return new mock_processor(kb_path);
+    }
+    else {
+      return new null_processor();
+    }
+  }
+
+
+}
 km_kbp_status
-km_kbp_keyboard_load(km_kbp_path_name kb_path,
-                                   km_kbp_keyboard **keyboard)
+km_kbp_keyboard_load(km_kbp_path_name kb_path, km_kbp_keyboard **keyboard)
 {
   assert(keyboard);
   if (!keyboard)
     return KM_KBP_STATUS_INVALID_ARGUMENT;
 
-  //auto stat = std::filesystem::status(kb_path);
-  //
-  // if (stat.type() != std::filesystem::file_type::regular)
-  //   return KM_KBP_STATUS_INVALID_ARGUMENT;
-
   try
   {
-    km::kbp::keyboard *kp = new km::kbp::keyboard(kb_path);
-    km_kbp_status status = kp->processor().validate();
+    abstract_processor *kp = processor_factory(kb_path);
+    km_kbp_status status = kp->validate();
     if (status != KM_KBP_STATUS_OK) {
       delete kp;
       return status;
     }
     *keyboard = static_cast<km_kbp_keyboard *>(kp);
   }
-  catch (std::bad_alloc) 
+  catch (std::bad_alloc)
   {
     return KM_KBP_STATUS_NO_MEM;
   }
@@ -69,6 +83,6 @@ km_kbp_keyboard_get_attrs(km_kbp_keyboard const *keyboard,
   if (!keyboard || !out)
     return KM_KBP_STATUS_INVALID_ARGUMENT;
 
-  *out = keyboard;
+  *out = &keyboard->keyboard();
   return KM_KBP_STATUS_OK;
 }