+++ linux/scripts/dist.sh
@@ -59,7 +59,7 @@ for proj in ${autotool_projects}; do
     rm -rf ../build-$proj
     mkdir -p ../build-$proj
     cd ../build-$proj
-    JSON_GLIB_CFLAGS="-I/usr/include/json-glib-1.0" JSON_GLIB_LIBS="-ljson-glib-1.0" KEYMAN_PROC_CFLAGS="-I/usr/include/keyman" KEYMAN_PROC_LIBS="-lkmnkbp" ../$proj/configure
+    GTK_CFLAGS="-I/usr/include/gtk-3.0" GTK_LIBS="-lgtk-3 -lgdk-3" JSON_GLIB_CFLAGS="-I/usr/include/json-glib-1.0" JSON_GLIB_LIBS="-ljson-glib-1.0" KEYMAN_PROC_CFLAGS="-I/usr/include/keyman" KEYMAN_PROC_LIBS="-lkmnkbp" ../$proj/configure
     make dist
     mv *.tar.gz ../dist
     cd $BASEDIR
@@ -81,7 +81,7 @@ for proj in ${extra_projects}; do
         datevers=`TZ=UTC git log -1 --pretty=format:%cd --date=format-local:%Y%m%d%H%M`
         kbpvers="keyman-keyboardprocessor-$vers~$datevers"
         cp -a keyboardprocessor $kbpvers
-        tar czf $kbpvers.tar.gz --exclude=debian\/ --exclude=build\/ --exclude=.gitignore $kbpvers
+        tar cvzf $kbpvers.tar.gz --exclude=debian --exclude=build --exclude=.gitignore $kbpvers
         rm -rf $kbpvers
         cp $kbpvers.tar.gz ../../linux/dist
     fi