+++ web/history.md
@@ -6,6 +6,7 @@
 * Fixes keyboard layout issues after mobile device rotations. (#248) (#970)
 * Fixes issue with oversized key text on some keyboards. (#382)
 * Fixes issue with diacritics not displaying on some keyboards.  (#1070)
+* Adds support for Promises to init() and setActiveKeyboard(). (#100)
 
 ## 2018-07-06 10.0.103 stable
 * Fixes issue for embedded Android, iOS apps where a keyboard with varying row counts in different layers could crash (#1055)